Rename .h to .hpp
Added README examples for TLRU and UTLRU
