--- inc/cappuccino/Cappuccino.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include "cappuccino/FifoCache.h"
-#include "cappuccino/LfuCache.h"
-#include "cappuccino/LfudaCache.h"
-#include "cappuccino/LruCache.h"
-#include "cappuccino/MruCache.h"
-#include "cappuccino/RrCache.h"
-#include "cappuccino/TlruCache.h"
-#include "cappuccino/UtlruCache.h"
