+++ inc/cappuccino/Cappuccino.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "cappuccino/FifoCache.hpp"
+#include "cappuccino/LfuCache.hpp"
+#include "cappuccino/LfudaCache.hpp"
+#include "cappuccino/LruCache.hpp"
+#include "cappuccino/MruCache.hpp"
+#include "cappuccino/RrCache.hpp"
+#include "cappuccino/TlruCache.hpp"
+#include "cappuccino/UtlruCache.hpp"
